# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Medication API",
      "description": "API for managing medications in the drug list.",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:3000",
        "description": "Local server"
      },
      {
        "url":"https://pharmacyapi-xz7h.onrender.com",
        "description": "Production Server"
      }
    ],
    "paths": {
      "/medication": {
        "get": {
          "summary": "Get all medications",
          "description": "Retrieve a list of all medications in the drug list.",
          "responses": {
            "200": {
              "description": "A list of medications",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "description": "The ID of the medication"
                        },
                        "brand_name": {
                          "type": "string",
                          "description": "The brand name of the medication"
                        },
                        "generic_name": {
                          "type": "string",
                          "description": "The generic name of the medication"
                        },
                        "manufacturer": {
                          "type": "string",
                          "description": "The manufacturer of the medication"
                        },
                        "dosage_strengths": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "List of dosage strengths"
                        },
                        "dosage_forms": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "List of dosage forms"
                        },
                        "drug_class": {
                          "type": "string",
                          "description": "The drug class of the medication"
                        }
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "description": "Error message"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new medication",
          "description": "Add a new medication to the drug list.",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "brand_name",
                    "generic_name",
                    "manufacturer",
                    "dosage_strengths",
                    "dosage_forms",
                    "drug_class"
                  ],
                  "properties": {
                    "brand_name": {
                      "type": "string"
                    },
                    "generic_name": {
                      "type": "string"
                    },
                    "manufacturer": {
                      "type": "string"
                    },
                    "dosage_strengths": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "dosage_forms": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "drug_class": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Medication created successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "The ID of the created medication"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "description": "Error message"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/medication/{id}": {
        "get": {
          "summary": "Get a medication by ID",
          "description": "Retrieve a specific medication using its ID.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "The ID of the medication",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "A specific medication",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "description": "The ID of the medication"
                      },
                      "brand_name": {
                        "type": "string",
                        "description": "The brand name of the medication"
                      },
                      "generic_name": {
                        "type": "string",
                        "description": "The generic name of the medication"
                      },
                      "manufacturer": {
                        "type": "string",
                        "description": "The manufacturer of the medication"
                      },
                      "dosage_strengths": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        },
                        "description": "List of dosage strengths"
                      },
                      "dosage_forms": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        },
                        "description": "List of dosage forms"
                      },
                      "drug_class": {
                        "type": "string",
                        "description": "The drug class of the medication"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Medication not found",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "description": "Error message"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "description": "Error message"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "put": {
          "summary": "Update a medication by ID",
          "description": "Update the details of an existing medication.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "The ID of the medication to update",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "brand_name",
                    "generic_name",
                    "manufacturer",
                    "dosage_strengths",
                    "dosage_forms",
                    "drug_class"
                  ],
                  "properties": {
                    "brand_name": {
                      "type": "string"
                    },
                    "generic_name": {
                      "type": "string"
                    },
                    "manufacturer": {
                      "type": "string"
                    },
                    "dosage_strengths": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "dosage_forms": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "drug_class": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Medication updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "The ID of the updated medication"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Medication not found or no changes made",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "description": "Error message"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "description": "Error message"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "delete": {
          "summary": "Delete a medication by ID",
          "description": "Delete a medication from the drug list by its ID.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "The ID of the medication to delete",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Medication deleted successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "description": "Success message"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid ID format",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "description": "Error message"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Medication not found",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "description": "Error message"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "description": "Error message"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
  